# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tinder Laravel API",
        "description": "API documentation for Tinder-like app",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Localhost API Server"
        }
    ],
    "paths": {
        "/api/like": {
            "post": {
                "tags": [
                    "Likes"
                ],
                "summary": "Like a user",
                "operationId": "8b479b1b5c3d3be75b6093800e25fafa",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "liked_user_id"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "liked_user_id": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User liked successfully"
                    }
                }
            }
        },
        "/api/dislike": {
            "post": {
                "tags": [
                    "Likes"
                ],
                "summary": "Dislike a user",
                "operationId": "03c26ab9b1b949c1d3db9a8343c71b5f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "liked_user_id"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "liked_user_id": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User disliked successfully"
                    }
                }
            }
        },
        "/api/liked": {
            "get": {
                "tags": [
                    "Likes"
                ],
                "summary": "Get list of users you liked",
                "operationId": "81b3caa7832863f92fc42cdf8d854d28",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of liked users"
                    }
                }
            }
        },
        "/api/recommendations": {
            "get": {
                "tags": [
                    "Recommendations"
                ],
                "summary": "Get recommended users (not yet liked/disliked)",
                "operationId": "f0ac77661847d67e935d73ea3c9000f0",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "location",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "min_age",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "max_age",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated list of recommended users"
                    }
                }
            }
        },
        "/api/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get all users",
                "operationId": "c457726701591d1183b53aa71fc13441",
                "responses": {
                    "200": {
                        "description": "List of all users"
                    }
                }
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Create new user",
                "operationId": "592819a0265360b2014512d6dbfaf0e7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "age",
                                    "location"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Alice"
                                    },
                                    "age": {
                                        "type": "integer",
                                        "example": 25
                                    },
                                    "pictures": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "pic1.jpg",
                                            "pic2.jpg"
                                        ]
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "New York"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User created successfully"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Likes",
            "description": "Like and Dislike actions"
        },
        {
            "name": "Recommendations",
            "description": "Recommended people list"
        },
        {
            "name": "Users",
            "description": "Users"
        }
    ]
}